Split dummy implementations into separate files
