+++ platform-controller/src/test/java/org/hobbit/controller/mocks/DummyResourceInformationCollector.java
@@ -1,0 +1,19 @@
+package org.hobbit.controller.mocks;
+
+import com.spotify.docker.client.messages.swarm.Task;
+import org.hobbit.controller.docker.ResourceInformationCollector;
+import org.hobbit.core.data.usage.ResourceUsageInformation;
+
+public class DummyResourceInformationCollector implements ResourceInformationCollector {
+
+    @Override
+    public ResourceUsageInformation getSystemUsageInformation() {
+        return null;
+    };
+
+    @Override
+    public ResourceUsageInformation getUsageInformation(Task.Criteria criteria) {
+        return null;
+    };
+
+}
